project-assignment/requirements.txt
fastapi
uvicorn[standard]
redis
motor
passlib[bcrypt]
phonenumbers
email-validator
python-dotenv
